New "comments" for our posts.
all tests passing.  Note there is a "comments_spec.rb" and a "byron_comments_spec.rb" - byron's is for creating the real spec tests, the other is just a source of samples, intended to be deleted later.

Note that there is a lot of code commented in routes and comments_spec

Co-authored-by: ellingtonjp <ellingtonjp@users.noreply.github.com>