update test config to render exceptions rather than raise them
by default in the dev and prod environments, rails will render an error
page and set an appropriate status code if a controller raises an
exception. however in the test environment, the exception is not caught.

to better simulate the dev/prod environment, we've changed the test
configuration to render error pages and response codes similar to the
dev and prod environments.