Get a bit more explicit with a test / code
Our code was written to throw exceptions and for
them to be handled by the rails router code, to
wrap them in appropriate HTTP responses - error
status codes and such.

Now, we have switched to a paradigm of being
more explicit, handling the exceptions in-line and
rendering the appropriate HTTP response.

Also made some miscellaneous config changes
to RSpec.

Co-authored-by: ellingtonjp <ellingtonjp@users.noreply.github.com>